Rename newsaggregator.types to news-aggregator.types and update imports across the application
