+++ src/models/news-aggregator.types.ts
